task-1/requirements.txt
requests==2.32.5
kafka-python==2.3.0
